feat: start adding logging
Also fix a bug where number of nodes in de-serialized SendNodes could
differ from what the number of nodes that is included in the packet is
supposed to be. With the fix, de-serializing fails if number of nodes
de-serialized differs from the number of nodes that bytes were supposed
to contain.