feat(dht): add `Kbucket` structure for holding lists of close nodes
It will hold lists of close nodes to own PK, and only 1 list of closest
nodes to friend's PK.

Add missing test for Bucket::try_add(), and fix a bug in this method,
where a node could not be added unless bucket was empty or node was
closer than any other node in the bucket.

Also modify `Bucket`, to work better with `Kbucket`.