feat: add methods to `Bucket` and `Kbucktes`
This makes them actually usable.

BREAKING CHANGE: `Bucket` no longer contains associated PK. Now it will
either have to be contained by the `Kbuckets` or `Bucket` should be
stored in a tuple with an associated PK. The alternative is to create a
struct for holding single bucket for nodes close to friends.