feat: add method to `DhtNode` for sending `SendNode`s
Also changed letter in Kbucket struct/methods to be consistent with what
it signifies in the spec. Previously `k` was used, which in the spec
represents number of nodes stored in a single Bucket, and was changed to
`n` which in the spec represents number of `Bucket`s in a Kbucket.